+++ src/ExtendedComparable.php
@@ -12,40 +12,40 @@ interface ExtendedComparable
      *
      * @return bool True if $this == $value.
      */
-    public function isEqualTo($value);
+    public function isEqualTo($value): bool;
 
     /**
      * @param mixed $value The value to compare.
      *
      * @return bool True if $this != $value.
      */
-    public function isNotEqualTo($value);
+    public function isNotEqualTo($value): bool;
 
     /**
      * @param mixed $value The value to compare.
      *
      * @return bool True if $this < $value.
      */
-    public function isLessThan($value);
+    public function isLessThan($value): bool;
 
     /**
      * @param mixed $value The value to compare.
      *
      * @return bool True if $this > $value.
      */
-    public function isGreaterThan($value);
+    public function isGreaterThan($value): bool;
 
     /**
      * @param mixed $value The value to compare.
      *
      * @return bool True if $this <= $value.
      */
-    public function isLessThanOrEqualTo($value);
+    public function isLessThanOrEqualTo($value): bool;
 
     /**
      * @param mixed $value The value to compare.
      *
      * @return bool True if $this >= $value.
      */
-    public function isGreaterThanOrEqualTo($value);
+    public function isGreaterThanOrEqualTo($value): bool;
 }