+++ src/SelfComparable.php
@@ -24,5 +24,5 @@ interface SelfComparable
      * @return int                              The result of the comparison.
      * @throws Exception\NotComparableException if $value is not the same type as $this.
      */
-    public function compare($value);
+    public function compare($value): int;
 }