+++ test/src/Comparator/SubClassComparableImpl.php
@@ -6,7 +6,7 @@ use Icecave\Parity\SubClassComparable;
 
 class SubClassComparableImpl implements SubClassComparable
 {
-    public function compare($value)
+    public function compare($value): int
     {
         return -10;
     }